extension by single-section example
